hot-order-agent/requirements.txt

streamlit
pandas
python-dotenv
openai
pyrfc==3.3.1
